+++ src/app/(main)/admin/member/_components/approve/index.ts
@@ -1,0 +1,2 @@
+export { ApproveMemberForm } from './Form'
+export { ApproveMemberTable } from './Table'
