+++ src/app/(main)/admin/member/_components/change-role/index.ts
@@ -1,0 +1,4 @@
+export { ChangeRoleDialog } from './Dialog'
+export { ChangeRoleDialogForm } from './DialogForm'
+export { RoleRadioGroup } from './RoleRadioGroup'
+export { ChangeRoleTable } from './Table'
